+++ api/src/main/java/jakarta/data/page/KeyPageable.java
@@ -1,0 +1,2 @@
+package jakarta.data.page;public interface KeyPageable {
+}
